Split solution into test app and exchange rate provider library
