--- jobs/Backend/Task/ExchangeRate.cs
@@ -1,23 +1,0 @@
-namespace ExchangeRateUpdater
-{
-    public class ExchangeRate
-    {
-        public ExchangeRate(Currency sourceCurrency, Currency targetCurrency, decimal value)
-        {
-            SourceCurrency = sourceCurrency;
-            TargetCurrency = targetCurrency;
-            Value = value;
-        }
-
-        public Currency SourceCurrency { get; }
-
-        public Currency TargetCurrency { get; }
-
-        public decimal Value { get; }
-
-        public override string ToString()
-        {
-            return $"{SourceCurrency}/{TargetCurrency}={Value}";
-        }
-    }
-}
