+++ jobs/Backend/Task/ExchangeRateProvider.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34728.123
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateProvider", "src\ExchangeRateProvider\ExchangeRateProvider.csproj", "{6E0AE99F-F7A3-4204-9C26-937721822017}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestApp", "src\TestApp\TestApp.csproj", "{31F4BEDF-F111-4947-A4CF-BA3DAE480602}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6E0AE99F-F7A3-4204-9C26-937721822017}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E0AE99F-F7A3-4204-9C26-937721822017}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E0AE99F-F7A3-4204-9C26-937721822017}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E0AE99F-F7A3-4204-9C26-937721822017}.Release|Any CPU.Build.0 = Release|Any CPU
+		{31F4BEDF-F111-4947-A4CF-BA3DAE480602}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{31F4BEDF-F111-4947-A4CF-BA3DAE480602}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{31F4BEDF-F111-4947-A4CF-BA3DAE480602}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{31F4BEDF-F111-4947-A4CF-BA3DAE480602}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7A85F406-5C5B-4BB5-A59F-5FD99C662A19}
+	EndGlobalSection
+EndGlobal
