+++ jobs/Backend/Task/src/ExchangeRateProvider/Currency.cs
@@ -1,0 +1,19 @@
+namespace ExchangeRateUpdater;
+
+public class Currency
+{
+    public Currency(string code)
+    {
+        Code = code;
+    }
+
+    /// <summary>
+    /// Three-letter ISO 4217 code of the currency.
+    /// </summary>
+    public string Code { get; }
+
+    public override string ToString()
+    {
+        return Code;
+    }
+}
