+++ jobs/Backend/Task/src/ExchangeRateProvider/ExchangeRatesDailyResponse.cs
@@ -1,0 +1,15 @@
+﻿public class CurrencyRate
+{
+	public string? ValidFor { get; set; }
+	public int? Order { get; set; }
+	public string? Country { get; set; }
+	public string? Currency { get; set; }
+	public int? Amount { get; set; }
+	public string? CurrencyCode { get; set; }
+	public decimal? Rate { get; set; }
+}
+
+public class ExchangeRatesDailyResponse
+{
+	public IEnumerable<CurrencyRate>? Rates { get; set; }
+}
